Last Fix

+++ Core/OtoKiralama.Application/Services/ReservationService.cs
@@ -1,7 +1,6 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
-using OtoKiralama.Application.Dtos.Car;
 using OtoKiralama.Application.Dtos.Pagination;
 using OtoKiralama.Application.Dtos.Reservation;
 using OtoKiralama.Application.Exceptions;
@@ -25,29 +24,63 @@ namespace OtoKiralama.Application.Services
             _userManager = userManager;
         }
 
+        public async Task CancelReservation(int id)
+        {
+            var existReservation = await _unitOfWork.ReservationRepository.GetEntity(r => r.Id == id);
+            if (existReservation is null)
+                throw new CustomException(404, "Id", "Reservation not found with this Id");
+            if (existReservation.Status == Domain.Enums.ReservationStatus.Canceled)
+                throw new CustomException(400, "Id", "Cannot cancel a reservation that is Cancelled");
+            existReservation.IsCanceled = true;
+            var existCar = await _unitOfWork.CarRepository.GetEntity(c => c.Id == existReservation.CarId);
+            existCar.IsReserved = false;
+            _unitOfWork.Commit();
+        }
+
+        public async Task CompleteReservation(int id)
+        {
+            var existReservation = await _unitOfWork.ReservationRepository.GetEntity(r => r.Id == id);
+            if (existReservation is null)
+                throw new CustomException(404, "Id", "Reservation not found with this Id");
+            if (existReservation.Status != Domain.Enums.ReservationStatus.Canceled)
+                throw new CustomException(400, "Id", "Cannot complete a reservation that is Cancelled");
+            existReservation.Status = Domain.Enums.ReservationStatus.Completed;
+            var existCar = await _unitOfWork.CarRepository.GetEntity(c => c.Id == existReservation.CarId);
+            existCar.IsReserved = false;
+            _unitOfWork.Commit();
+        }
+
         public async Task CreateReservationAsync(ReservationCreateDto reservationCreateDto)
         {
             var existCar = await _unitOfWork.CarRepository.isExists(c => c.Id == reservationCreateDto.CarId);
             if (!existCar)
                 throw new CustomException(404, "CarId", "Car not found with this Id");
+            var existCarEntity = await _unitOfWork.CarRepository.GetEntity(c => c.Id == reservationCreateDto.CarId);
+            if (existCarEntity.IsReserved)
+                throw new CustomException(400, "CarId", "Car is already reserved");
+            if (!existCarEntity.IsActive)
+                throw new CustomException(400, "CarId", "Car is not active");
             var existUser = await _userManager.FindByIdAsync(reservationCreateDto.UserId);
             if (existUser is null)
                 throw new CustomException(404, "UserId", "User not found with this Id");
-            bool isCarAvailable = await IsCarAvailable(reservationCreateDto.CarId, reservationCreateDto.StartDate, reservationCreateDto.EndDate);
-            if (!isCarAvailable)
-                throw new CustomException(400, "CarId", "Car is not available for this date range");
             var reservation = _mapper.Map<Reservation>(reservationCreateDto);
             reservation.AppUserId = reservationCreateDto.UserId;
-            reservation.TotalPrice = (reservation.EndDate - reservation.StartDate).TotalDays * reservation.Car.DailyPrice;
+            reservation.TotalPrice = (reservation.EndDate - reservation.StartDate).TotalDays * existCarEntity.DailyPrice;
+            reservation.Status = Domain.Enums.ReservationStatus.Pending;
+            existCarEntity.IsReserved = true;
             await _unitOfWork.ReservationRepository.Create(reservation);
             _unitOfWork.Commit();
         }
 
         public async Task DeleteReservationAsync(int id)
         {
-            var existReservation = await _unitOfWork.ReservationRepository.GetEntity(r=>r.Id==id);
+            var existReservation = await _unitOfWork.ReservationRepository.GetEntity(r => r.Id == id);
             if (existReservation is null)
                 throw new CustomException(400, "Id", "Reservation with this Id is not found");
+
+            if (existReservation.Status == Domain.Enums.ReservationStatus.InProgress)
+                throw new CustomException(400, "Id", "Cannot delete a reservation that is in progress");
+
             await _unitOfWork.ReservationRepository.Delete(existReservation);
             _unitOfWork.Commit();
         }
@@ -61,7 +94,7 @@ namespace OtoKiralama.Application.Services
                         .ThenInclude(c => c.Brand)
                     .Include(c => c.Car)
                         .ThenInclude(c => c.Model)
-                            .ThenInclude(m=> m.CarPhoto)
+                            .ThenInclude(m => m.CarPhoto)
                     .Include(c => c.Car)
                         .ThenInclude(c => c.Body)
                     .Include(c => c.Car)
@@ -117,18 +150,6 @@ namespace OtoKiralama.Application.Services
                 throw new CustomException(404, "Id", "Reservation not found with this Id");
 
             return _mapper.Map<ReservationReturnDto>(reservation);
-        }
-
-        public async Task<bool> IsCarAvailable(int carId, DateTime startDate, DateTime endDate)
-        {
-            var reservations = await _unitOfWork.ReservationRepository.GetAll(
-                predicate: r => r.CarId == carId &&
-                              ((r.StartDate >= startDate && r.StartDate <= endDate) ||
-                               (r.EndDate >= startDate && r.EndDate <= endDate) ||
-                               (r.StartDate <= startDate && r.EndDate >= endDate))
-            );
-
-            return reservations.Count == 0;
         }
     }
 }
